fix pool size to size_t

+++ src/allocator.h
@@ -7,11 +7,11 @@ static const char* layout_name = "hashtable";
 struct Allocator {
  public:
 #ifdef PMEM
-  static void Initialize(const char* pool_name, uint32_t pool_size) {
+  static void Initialize(const char* pool_name, size_t pool_size) {
     instance_ = new Allocator(pool_name, pool_size);
   }
 
-  Allocator(const char* pool_name, uint32_t pool_size) {
+  Allocator(const char* pool_name, size_t pool_size) {
     if (!FileExists(pool_name)) {
       LOG("creating a new pool");
       pm_pool_ =
